app/pyproject.toml
[tool.poetry]
name = "princetonpy"
version = "0.1.0"
description = "A website for the Princeton Python users group."
authors = ["mhadam <michael@hadam.us>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
django-ses = {extras = ["events"], version = "^2.4.0"}
django-newsletter = "^0.9.1"
django-tinymce = "^3.2.0"
sorl-thumbnail = "^12.7.0"
django-environ = "^0.8.0"
yacron = "^0.14.0"
whitenoise = {extras = ["brotli"], version = "^6.0.0"}
Pillow = "^9.3.0"
hypercorn = "^0.14.3"

[tool.poetry.dev-dependencies]
pytest-django = "^4.4.0"
Faker = "^13.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
